Move payement rules is separate table
- revert member type  edition page
- create new  menu etnry and payrule edition page
- extend texual description of payrule definition
